requirements.txt
django==4.2.7
djangorestframework==3.14.0
langchain==0.0.335
faiss-cpu==1.7.4
pymupdf==1.23.6
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
chromadb==0.4.18
pdfplumber==0.10.2
gunicorn==21.2.0
whitenoise==6.6.0 
langchain-community==0.2.10
langchain-core==0.3.10
